% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{auc
model auc}
\usage{
auc(response, predictor, od, ...)
}
\arguments{
\item{response}{response value}

\item{predictor}{predictor value}

\item{od}{output dirname}

\item{...}{other parameter}
}
\value{
roc object
}
\description{
auc
model auc
}
