% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinAndExpMerge.R
\name{ClinAndExpMerge}
\alias{ClinAndExpMerge}
\title{MainRun
find signature and survival analysis}
\usage{
ClinAndExpMerge(expMat, clinData, surv.col = c("status", "OS"), ...)
}
\arguments{
\item{expMat}{survival and expression data, col is status,time,g1,g2,...}

\item{clinData}{clinical survival data, must contain two columns: survival status and survival time}

\item{surv.col}{columns of status and time, default c("status", "OS")}

\item{...}{other parameter}
}
\value{
data.frame, Contains survival and expression data
}
\description{
MainRun
find signature and survival analysis
}
