% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_data_process.R
\name{cox_data_process}
\alias{cox_data_process}
\title{cox_data_process
data processing
cox input data}
\usage{
cox_data_process(data, var.names)
}
\arguments{
\item{data}{row is sample, col is phelotype variate,example age and AJCC stage}

\item{var.names}{phelotype variate}
}
\value{
data.frame
}
\description{
cox_data_process
data processing
cox input data
}
