% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpDataProcess.R
\name{ExpDataProcess}
\alias{ExpDataProcess}
\title{ExpDataProcess
convert gene symbol and merge data}
\usage{
ExpDataProcess(data1, data2 = NULL, id.col = NULL, ...)
}
\arguments{
\item{data1}{data.frame,geneID and gene expression}

\item{data2}{data.frame, gene information}

\item{id.col}{set column, example:1,2 or "ID,symbol"}

\item{...}{other parameter}
}
\value{
expression data.frame
}
\description{
ExpDataProcess
convert gene symbol and merge data
}
