% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcClinOS.R
\name{gdcClinOS}
\alias{gdcClinOS}
\title{gdcClinOS
get overall survival information form GDC}
\usage{
gdcClinOS(
  mfile,
  mjson,
  col.n = c("entity_submitter_id", "days_to_death", "days_to_last_follow_up",
    "vital_status")
)
}
\arguments{
\item{mfile}{GDC clinical.cases_selection file}

\item{mjson}{GDC metadata.cart json}

\item{col.n}{select mfile colnames ,defaut:c("entity_submitter_id", "days_to_death", "days_to_last_follow_up", "vital_status")}

\item{...}{other parameter}
}
\description{
gdcClinOS
get overall survival information form GDC
}
