% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcExpression.R
\name{gdcExp}
\alias{gdcExp}
\title{gdcExp
get expression matrix form GDC}
\usage{
gdcExp(idir, mjson)
}
\arguments{
\item{idir}{GDC case expreesion path}

\item{mjson}{GDC metadata.cart json}

\item{...}{other parameter}
}
\description{
gdcExp
get expression matrix form GDC
}
