% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcSampleID.R
\name{gdcSampleID}
\alias{gdcSampleID}
\title{gdcSampleID
Mapping between file names and sample names}
\usage{
gdcSampleID(mjson)
}
\arguments{
\item{mjson}{GDC metadata.cart json}

\item{...}{other parameter}
}
\description{
gdcSampleID
Mapping between file names and sample names
}
