% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDataFilter.R
\name{GetFilterExpression}
\alias{GetFilterExpression}
\title{GetFilterExpression
filter sample by gene expression percentage}
\usage{
GetFilterExpression(data, percent = 0.1)
}
\arguments{
\item{data}{TCGA expression data, row is gene,column is sample}

\item{percent}{gene expression percentage
default 0.1}

\item{...}{other parameters}
}
\value{
expression matrix
}
\description{
GetFilterExpression
filter sample by gene expression percentage
}
