% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOClinical.R
\name{getGEOclinical}
\alias{getGEOclinical}
\title{getGEOclinical
get clinical information form GEO}
\usage{
getGEOclinical(mfile)
}
\arguments{
\item{mfile}{GEO series_matrix.txt}

\item{...}{other parameter}
}
\description{
getGEOclinical
get clinical information form GEO
}
