% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOProbeID.R
\name{getGEOProbeID}
\alias{getGEOProbeID}
\title{getGEOProbeID
get gene information form GEO}
\usage{
getGEOProbeID(mfile)
}
\arguments{
\item{mfile}{GEO family.soft}

\item{...}{other parameter}
}
\description{
getGEOProbeID
get gene information form GEO
}
