% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTumorID.R
\name{getTumorID}
\alias{getTumorID}
\title{getTumorID: tumor sample ID list
TCGA sample name, "-" split}
\usage{
getTumorID(sample, ...)
}
\arguments{
\item{sample}{is vector}

\item{...}{other parameters}
}
\value{
a vector
}
\description{
getTumorID: tumor sample ID list
TCGA sample name, "-" split
}
