% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSurvival.R
\name{KMSurvival}
\alias{KMSurvival}
\title{KMSurvival 
Run Kaplan-Meier survival analysis for biomarker}
\usage{
KMSurvival(object, var.name, od, xname = "OS", ...)
}
\arguments{
\item{object}{survival data,first column is status,second column is time,...}

\item{var.name}{colnames for survfit}

\item{od}{output dirname}

\item{...}{other parameter}
}
\value{
data.frame
}
\description{
KMSurvival 
Run Kaplan-Meier survival analysis for biomarker
}
