% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFeatureChoose.R
\name{MainFeatureChoose}
\alias{MainFeatureChoose}
\title{MainFeatureChoose
Prognostic markers}
\usage{
MainFeatureChoose(data, geneS, od)
}
\arguments{
\item{data}{clinical and expression data}

\item{geneS}{candidate gene set}

\item{od}{output dirname}

\item{...}{other parameters}
}
\value{
data.frame
}
\description{
MainFeatureChoose
Prognostic markers
}
