% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainRiskSurv.R
\name{MainRiskSurv}
\alias{MainRiskSurv}
\title{MainRiskSurv}
\usage{
MainRiskSurv(data, choose_feature, od)
}
\arguments{
\item{data}{clinical and expression data}

\item{choose_feature}{signature gene set}

\item{od}{output dirname}

\item{...}{other parameters}
}
\value{
data.frame
}
\description{
define risk group and survival for choose feature
}
