% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainRun.R
\name{MainRun}
\alias{MainRun}
\title{MainRun
find signature and survival analysis}
\usage{
MainRun(
  initiData,
  geneS,
  out,
  pvalcut = 0.05,
  key = "OS",
  phenoData = NULL,
  var.pheno = NULL,
  ...
)
}
\arguments{
\item{initiData}{survival status/time and expression data,column is status,time,g1,g2,...,gn}

\item{geneS}{vector,genes}

\item{out}{output dirname}

\item{pvalcut}{unicox model pvalue, filter gene, default 0.05}

\item{key}{OS or DFS
default OS}

\item{phenoData}{phelotype dataset, rowname is sample}

\item{var.pheno}{vector phelotype name,example:age,sex,AJCC}

\item{...}{other parameter}
}
\description{
MainRun
find signature and survival analysis
}
