% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeFeature.R
\name{OptimizeFeature}
\alias{OptimizeFeature}
\title{final feature
use univariate COX and LASSO regression optimize feature}
\usage{
OptimizeFeature(data, pvalcut = 0.05, od, n = 1000, ...)
}
\arguments{
\item{data}{clinical and expression data, fist column is OS status; second column is OS time; feature expression...}

\item{pvalcut}{univariate cox pvalue cutoff
default pvalcut is 0.05}

\item{od}{output dirname}

\item{n}{random times
default 1000}

\item{...}{other parameter}
}
\value{
data.frame genes and weight
}
\description{
Random Perturbation
description: random n(1000) times OptimizeFeature, return gene set and gene weight
}
