% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurvivalCurve.R
\name{PlotSurvivalCurve}
\alias{PlotSurvivalCurve}
\title{PlotSurvivalCurve}
\usage{
PlotSurvivalCurve(survival_plot, dir, xname = "OS", title = NULL)
}
\arguments{
\item{survival_plot}{survfit object}

\item{dir}{output dirname}

\item{xname}{xlab name
default OS}

\item{title}{title}
}
\value{
p,plot object
}
\description{
plot survival-curve picture
}
