% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score.R
\name{RiskScore}
\alias{RiskScore}
\title{RiskScore}
\usage{
RiskScore(data, feature, method = "median")
}
\arguments{
\item{data}{TCGA expression data}

\item{feature}{choose feature,example g1,g2,g3}

\item{...}{other parameters}
}
\value{
data.frame
}
\description{
calculate prognostic index(PI) or risk score
risk_score = Beta(1)Exp(1)+Beta(2)Exp(2)+...+Beta(p)Exp(p)
}
