% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGeneSurv.R
\name{singleGS}
\alias{singleGS}
\title{singleGS}
\usage{
singleGS(
  data,
  gene,
  out,
  method = c("quartile", "median"),
  high.quantile = 2/3
)
}
\arguments{
\item{data}{clinical ana expresion data.frame}

\item{gene}{g name}

\item{out}{output dirname}

\item{method}{"quartile" or "median", use defined high/low}

\item{high.quantile}{if methos = "quartile", the quantile of high group}
}
\value{

}
\description{
plot single g survival analysis,high:T3 ,low/in:T1-2
}
