% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StableModel.R
\name{StableModel}
\alias{StableModel}
\title{StableModel
select gene that reached the highest C-index to the signature}
\usage{
StableModel(Data, covariates, seed = NULL, od)
}
\arguments{
\item{Data}{survival status/time and expression data,column is status,time,g1,g2,...,gn}

\item{covariates}{vector,genes}

\item{od}{output dirname}

\item{...}{other parameter}
}
\description{
StableModel
select gene that reached the highest C-index to the signature
}
