% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniCox.R
\name{uniCox}
\alias{uniCox}
\title{Cox proportional hazards regression analysis}
\usage{
uniCox(data, var = NULL, method = "univariate")
}
\arguments{
\item{data}{row is TCGA sample,col is variable}

\item{var}{variable name,example sex,age,...}

\item{method:}{univariate or multivariate}

\item{...}{other parameter}
}
\description{
Cox proportional hazards regression analysis
}
